unix_requirements.txt
anyio==3.6.2
APScheduler==3.10.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
Authlib==1.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
cryptography==40.0.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.7
django-crontab==0.7.1
django-oauth-toolkit==2.2.0
django-polymorphic==3.1.0
django-tagify2==1.0.19
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fqdn==1.5.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jwcrypto==1.4.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.3
oauth==1.0.1
oauthlib==3.2.2
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
prometheus-client==0.16.0
prompt-toolkit==3.0.37
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pypng==0.20220715.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
qrcode==7.4.2
qtconsole==5.4.0
QtPy==2.3.0
reportlab==3.6.12
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
selenium==4.8.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
stripe==5.4.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
Unidecode==1.3.6
uri-template==1.2.0
urllib3==1.26.15
virtualenv==20.16.6
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
wrapt==1.15.0
wsproto==1.2.0
zipp==3.14.0
